src/save2win-engine/requirements.txt
Flask==2.2.2
requests==2.28.1
gunicorn==20.1.0
werkzeug==2.2.2
google-cloud-aiplatform>=1.38.0
pyjwt 
cryptography